Feat: HomeEntireImageStore 분리
